toolkit/components/extensions/test/mochitest/test_ext_simple.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for simple WebExtension</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SpawnTask.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/ExtensionTestUtils.js"></script>
  <script type="text/javascript" src="head.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>

<script type="application/javascript;version=1.8">

add_task(function* test_simple() {
  let extension = ExtensionTestUtils.loadExtension("simple");
  info("load complete");
  yield extension.startup();
  info("startup complete");
  yield extension.unload();
  info("extension unloaded successfully");
});

add_task(function* test_background() {
  let extension = ExtensionTestUtils.loadExtension("background");
  info("load complete");
  let [, x] = yield Promise.all([extension.startup(), extension.awaitMessage("running")]);
  is(x, 1, "got correct value from extension");
  info("startup complete");
  extension.sendMessage(10, 20);
  yield extension.awaitFinish();
  info("test complete");
  yield extension.unload();
  info("extension unloaded successfully");
});

</script>

</body>
</html>
